requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
base58==2.1.1
bitarray==2.7.3
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cytoolz==0.12.1
easyocr==1.4
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
Faker==18.9.0
filelock==3.12.0
Flask==2.1.1
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
imageio==2.29.0
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
lazy_loader==0.2
lru-dict==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
parsimonious==0.8.1
Pillow==9.5.0
protobuf==3.20.3
pycryptodome==3.18.0
pyrsistent==0.19.3
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.1
PyWavelets==1.4.1
PyYAML==6.0
random-username==1.0.2
requests==2.31.0
rlp==2.0.1
scikit-image==0.20.0
scipy==1.10.1
six==1.16.0
sympy==1.12
tifffile==2023.4.12
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.6.2
urllib3==2.0.2
varint==1.0.2
web3==5.24.0
websockets==9.1
Werkzeug==2.3.4
yarl==1.9.2
